Use HandshaeResponse class insted of manual build json
