+++ CMakeLists.txt
@@ -68,12 +68,12 @@ if(FULLNODE)
     endif()
 endif()
 
-if(WITH_LIGHTNODE)
-    add_subdirectory(lightnode)
-endif()
-
 if(WITH_CPPSDK)
     add_subdirectory(bcos-sdk)
+endif()
+
+if(WITH_LIGHTNODE)
+    add_subdirectory(lightnode)
 endif()
 
 if(WITH_BENCHMARK)
